Code refactored.
