+++ SeriesCruncher/app/src/main/java/com/balazs_csernai/seriescruncher/seriesdetails/model/episode/EpisodeComparator.java
@@ -1,4 +1,4 @@
-package com.balazs_csernai.seriescruncher.seriesdetails.model;
+package com.balazs_csernai.seriescruncher.seriesdetails.model.episode;
 
 import java.util.Comparator;
