+++ SeriesCruncher/app/src/main/java/com/balazs_csernai/seriescruncher/seriesdetails/ui/SeriesDetailsScreen.java
@@ -1,8 +1,7 @@
 package com.balazs_csernai.seriescruncher.seriesdetails.ui;
 
-import com.balazs_csernai.seriescruncher.seriesdetails.model.EpisodeListModel;
-import com.balazs_csernai.seriescruncher.image.ImageLoader.ImageTarget;
-import com.balazs_csernai.seriescruncher.seriesdetails.model.PosterModel;
+import com.balazs_csernai.seriescruncher.seriesdetails.model.episode.EpisodeListModel;
+import com.balazs_csernai.seriescruncher.seriesdetails.model.poster.PosterModel;
 
 /**
  * Created by Erik_Markus_Kramli on 2016-01-13.
@@ -14,8 +13,6 @@ public interface SeriesDetailsScreen {
     void show(String title);
 
     void show(EpisodeListModel episodes);
-
-    ImageTarget getPosterImageTarget();
 
     void show(PosterModel posterModel);
 }
