+++ SeriesCruncher/app/src/main/java/com/balazs_csernai/seriescruncher/utils/converter/EpisodeListItemConverter.java
@@ -1,9 +1,9 @@
 package com.balazs_csernai.seriescruncher.utils.converter;
 
 
-import com.balazs_csernai.seriescruncher.seriesdetails.model.EpisodeModel;
-import com.balazs_csernai.seriescruncher.seriesdetails.model.EpisodeListItemEntity;
-import com.balazs_csernai.seriescruncher.seriesdetails.model.EpisodeListItemModel;
+import com.balazs_csernai.seriescruncher.seriesdetails.model.episode.EpisodeModel;
+import com.balazs_csernai.seriescruncher.seriesdetails.model.episode.EpisodeListItemEntity;
+import com.balazs_csernai.seriescruncher.seriesdetails.model.episode.EpisodeListItemModel;
 
 import javax.inject.Inject;
